# spec/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ory/keto/opl/v1alpha1/syntax_service.proto",
    "version": "version not set"
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/relation-tuples/check": {
      "get": {
        "summary": "Performs an authorization check.",
        "operationId": "CheckService_Check",
        "responses": {
          "200": {
            "description": "The response of the permission check, in case it is allowed. For the `/openapi` endpoints, the status code is always 200.",
            "schema": {
              "$ref": "#/definitions/checkPermissionResult"
            }
          },
          "403": {
            "description": "The response of the permission check, in case it is not allowed. For the `/openapi` endpoints, the status code is always 200.",
            "schema": {
              "$ref": "#/definitions/checkPermissionResult"
            }
          }
        },
        "parameters": [
          {
            "name": "namespace",
            "description": "The namespace to evaluate the check.\n\nNote: If you use the expand-API and the check\nevaluates a RelationTuple specifying a SubjectSet as\nsubject or due to a rewrite rule in a namespace config\nthis check request may involve other namespaces automatically.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "object",
            "description": "The related object in this check.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "relation",
            "description": "The relation between the Object and the Subject.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_id",
            "description": "A concrete id of the subject.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_set.namespace",
            "description": "The namespace of the object and relation\nreferenced in this subject set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_set.object",
            "description": "The object related by this subject set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_set.relation",
            "description": "The relation between the object and the subjects.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "max-depth",
            "description": "The maximum depth to search for a relation.\n\nIf the value is less than 1 or greater than the global\nmax-depth then the global max-depth will be used instead.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": ["permission"],
        "consumes": ["application/x-www-form-urlencoded", "application/json"]
      },
      "post": {
        "summary": "Performs an authorization check.",
        "operationId": "CheckService_Check3",
        "responses": {
          "200": {
            "description": "The response of the permission check, in case it is allowed. For the `/openapi` endpoints, the status code is always 200.",
            "schema": {
              "$ref": "#/definitions/checkPermissionResult"
            }
          },
          "403": {
            "description": "The response of the permission check, in case it is not allowed. For the `/openapi` endpoints, the status code is always 200.",
            "schema": {
              "$ref": "#/definitions/checkPermissionResult"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "The request for a CheckService.Check RPC.\nChecks whether a specific subject is related to an object.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/postCheckPermissionBody"
            }
          }
        ],
        "tags": ["permission"],
        "consumes": ["application/x-www-form-urlencoded", "application/json"]
      }
    },
    "/relation-tuples/check/openapi": {
      "get": {
        "summary": "Performs an authorization check.",
        "operationId": "CheckService_Check2",
        "responses": {
          "200": {
            "description": "The response of the permission check, in case it is allowed. For the `/openapi` endpoints, the status code is always 200.",
            "schema": {
              "$ref": "#/definitions/checkPermissionResult"
            }
          },
          "403": {
            "description": "The response of the permission check, in case it is not allowed. For the `/openapi` endpoints, the status code is always 200.",
            "schema": {
              "$ref": "#/definitions/checkPermissionResult"
            }
          }
        },
        "parameters": [
          {
            "name": "namespace",
            "description": "The namespace to evaluate the check.\n\nNote: If you use the expand-API and the check\nevaluates a RelationTuple specifying a SubjectSet as\nsubject or due to a rewrite rule in a namespace config\nthis check request may involve other namespaces automatically.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "object",
            "description": "The related object in this check.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "relation",
            "description": "The relation between the Object and the Subject.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_id",
            "description": "A concrete id of the subject.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_set.namespace",
            "description": "The namespace of the object and relation\nreferenced in this subject set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_set.object",
            "description": "The object related by this subject set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_set.relation",
            "description": "The relation between the object and the subjects.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "max-depth",
            "description": "The maximum depth to search for a relation.\n\nIf the value is less than 1 or greater than the global\nmax-depth then the global max-depth will be used instead.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": ["permission"],
        "consumes": ["application/x-www-form-urlencoded", "application/json"]
      },
      "post": {
        "summary": "Performs an authorization check.",
        "operationId": "CheckService_Check4",
        "responses": {
          "200": {
            "description": "The response of the permission check, in case it is allowed. For the `/openapi` endpoints, the status code is always 200.",
            "schema": {
              "$ref": "#/definitions/checkPermissionResult"
            }
          },
          "403": {
            "description": "The response of the permission check, in case it is not allowed. For the `/openapi` endpoints, the status code is always 200.",
            "schema": {
              "$ref": "#/definitions/checkPermissionResult"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "The request for a CheckService.Check RPC.\nChecks whether a specific subject is related to an object.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/postCheckPermissionBody"
            }
          }
        ],
        "tags": ["permission"],
        "consumes": ["application/x-www-form-urlencoded", "application/json"]
      }
    },
    "/relation-tuples/expand": {
      "get": {
        "summary": "Expands the subject set into a tree of subjects.",
        "operationId": "expandPermissions",
        "responses": {
          "200": {
            "description": "The response of the permission check, in case it is allowed. For the `/openapi` endpoints, the status code is always 200.",
            "schema": {
              "$ref": "#/definitions/expandedPermissionTree"
            }
          }
        },
        "parameters": [
          {
            "name": "max-depth",
            "description": "The maximum depth of tree to build.\n\nIf the value is less than 1 or greater than the global\nmax-depth then the global max-depth will be used instead.\n\nIt is important to set this parameter to a meaningful\nvalue. Ponder how deep you really want to display this.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "namespace",
            "description": "The namespace of the object and relation\nreferenced in this subject set.",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "object",
            "description": "The object related by this subject set.",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "relation",
            "description": "The relation between the object and the subjects.",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "tags": ["permission"],
        "consumes": ["application/x-www-form-urlencoded"]
      }
    },
...
  },
  "definitions": {
...
    "postCheckPermissionBody": {
      "type": "object",
      "properties": {
        "namespace": {
          "type": "string",
          "description": "The namespace to evaluate the check.\n\nNote: If you use the expand-API and the check\nevaluates a RelationTuple specifying a SubjectSet as\nsubject or due to a rewrite rule in a namespace config\nthis check request may involve other namespaces automatically."
        },
        "object": {
          "type": "string",
          "description": "The related object in this check."
        },
        "relation": {
          "type": "string",
          "description": "The relation between the Object and the Subject."
        },
        "subject_id": {
          "type": "string",
          "description": "A concrete id of the subject."
        },
        "subject_set": {
          "$ref": "#/definitions/subjectSetQuery",
          "description": "A subject set that expands to more Subjects.\nMore information are available under [concepts](../concepts/subjects.mdx)."
        },
        "max-depth": {
          "type": "integer",
          "format": "int32",
          "description": "The maximum depth to search for a relation.\n\nIf the value is less than 1 or greater than the global\nmax-depth then the global max-depth will be used instead."
        }
      },
      "description": "The request for a CheckService.Check RPC.\nChecks whether a specific subject is related to an object."
    },
    "checkPermissionResult": {
      "type": "object",
      "properties": {
        "allowed": {
          "type": "boolean",
          "description": "Whether the specified subject (id)\nis related to the requested object.\n\nIt is false by default if no ACL matches."
        }
      },
      "description": "The response for a CheckService.Check rpc."
    },
...
    "ory.keto.relation_tuples.v1alpha2.NodeType": {
      "type": "string",
      "enum": [
        "unspecified",
        "union",
        "exclusion",
        "intersection",
        "leaf",
        "tuple_to_subject_set",
        "computed_subject_set",
        "not"
      ],
      "default": "unspecified"
    },
    "relationship": {
      "type": "object",
      "properties": {
        "namespace": {
          "type": "string",
          "description": "The namespace this relation tuple lives in."
        },
        "object": {
          "type": "string",
          "description": "The object related by this tuple.\nIt is an object in the namespace of the tuple."
        },
        "relation": {
          "type": "string",
          "description": "The relation between an Object and a Subject."
        },
        "subject_id": {
          "type": "string",
          "description": "A concrete id of the subject."
        },
        "subject_set": {
          "$ref": "#/definitions/subjectSet",
          "description": "A subject set that expands to more Subjects.\nMore information are available under [concepts](../concepts/subjects.mdx)."
        }
      },
      "description": "RelationTuple defines a relation between an Object and a Subject.",
      "required": ["namespace", "object", "relation"]
    },
...
    "ory.keto.relation_tuples.v1alpha2.Subject": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "A concrete id of the subject."
        },
        "set": {
          "$ref": "#/definitions/subjectSet",
          "description": "A subject set that expands to more Subjects.\nMore information are available under [concepts](../concepts/subjects.mdx)."
        }
      },
      "description": "Subject is either a concrete subject id or\na `SubjectSet` expanding to more Subjects."
    },
    "subjectSet": {
      "type": "object",
      "properties": {
        "namespace": {
          "type": "string",
          "description": "The namespace of the object and relation\nreferenced in this subject set."
        },
        "object": {
          "type": "string",
          "description": "The object related by this subject set."
        },
        "relation": {
          "type": "string",
          "description": "The relation between the object and the subjects."
        }
      },
      "description": "SubjectSet refers to all subjects who have\nthe same `relation` on an `object`.",
      "required": ["namespace", "object", "relation"]
    },
    "subjectSetQuery": {
      "type": "object",
      "properties": {
        "namespace": {
          "type": "string",
          "description": "The namespace of the object and relation\nreferenced in this subject set."
        },
        "object": {
          "type": "string",
          "description": "The object related by this subject set."
        },
        "relation": {
          "type": "string",
          "description": "The relation between the object and the subjects."
        }
      },
      "description": "SubjectSetQuery refers to all subjects who have\nthe same `relation` on an `object`."
    },
    "expandedPermissionTree": {
      "type": "object",
      "properties": {
        "type": {
          "$ref": "#/definitions/ory.keto.relation_tuples.v1alpha2.NodeType",
          "description": "The type of the node."
        },
        "subject": {
          "$ref": "#/definitions/ory.keto.relation_tuples.v1alpha2.Subject",
          "description": "The subject this node represents.\nDeprecated: More information is now available in the tuple field."
        },
        "tuple": {
          "$ref": "#/definitions/relationship",
          "description": "The relation tuple this node represents."
        },
        "children": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/expandedPermissionTree"
          },
          "description": "The children of this node.\n\nThis is never set if `node_type` == `NODE_TYPE_LEAF`."
        }
      },
      "required": ["type"]
    },
...
  }
}